feat: Add image resizing script for StyleGAN
The code changes include adding a new script `make_resized_data.py` in the `StyleGAN` directory. This script resizes images from the `FFHQ/images1024x1024` directory to a size of 128x128 pixels and saves them in the `FFHQ_resized` directory. This script will be useful for preparing the data for training the StyleGAN model.